# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hotel Booking Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: propertyId
          in: query
          required: false
          schema:
            type: integer
        - name: unitId
          in: query
          required: false
          schema:
            type: integer
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /finance/monthly:
    get:
      operationId: getFinanceMonthly
      tags: [expenses]
      summary: Get monthly revenue vs expenses breakdown
      parameters:
        - name: propertyId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Monthly finance breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyFinance"
...
  /field-users:
    get:
      operationId: listFieldUsers
      tags: [fieldUsers]
      summary: List field operations users
      parameters:
        - name: propertyId
          in: query
          schema:
            type: integer
        - name: role
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of field users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FieldUser"
    post:
      operationId: createFieldUser
      tags: [fieldUsers]
      summary: Create a field user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FieldUserInput"
      responses:
        "201":
          description: Field user created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldUser"
  /field-users/{id}:
    patch:
      operationId: updateFieldUser
      tags: [fieldUsers]
      summary: Update a field user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FieldUserInput"
      responses:
        "200":
          description: Updated field user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldUser"
    delete:
      operationId: deleteFieldUser
      tags: [fieldUsers]
      summary: Delete a field user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/custom-fields:
    get:
      operationId: listCustomFields
      tags: [adminSettings]
      summary: List custom field definitions
      parameters:
        - name: entityType
          in: query
          schema:
            type: string
          description: "Filter by entity type: asset | task"
      responses:
        "200":
          description: List of custom fields
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomField"
    post:
      operationId: createCustomField
      tags: [adminSettings]
      summary: Create a new custom field definition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomFieldInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomField"
  /admin/custom-fields/{id}:
    patch:
      operationId: updateCustomField
      tags: [adminSettings]
      summary: Update a custom field definition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomFieldInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomField"
    delete:
      operationId: deleteCustomField
      tags: [adminSettings]
      summary: Delete a custom field definition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    CustomField:
      type: object
      required: [id, entityType, fieldKey, fieldLabel, fieldType, required, sortOrder, active, createdAt]
      properties:
        id:
          type: integer
        entityType:
          type: string
          description: "asset | task"
        fieldKey:
          type: string
          description: Machine-readable key (e.g. serial_number)
        fieldLabel:
          type: string
          description: Display label shown in the UI
        fieldType:
          type: string
          description: "text | number | select | date | boolean"
        options:
          type: array
          nullable: true
          items:
            type: string
          description: Available options for select-type fields
        required:
          type: boolean
        sortOrder:
          type: integer
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateCustomFieldInput:
      type: object
      required: [entityType, fieldKey, fieldLabel, fieldType]
      properties:
        entityType:
          type: string
        fieldKey:
          type: string
        fieldLabel:
          type: string
        fieldType:
          type: string
        options:
          type: array
          nullable: true
          items:
            type: string
        required:
          type: boolean
        sortOrder:
          type: integer
        active:
          type: boolean
    UpdateCustomFieldInput:
      type: object
      properties:
        fieldLabel:
          type: string
        fieldType:
          type: string
        options:
          type: array
          nullable: true
          items:
            type: string
        required:
          type: boolean
        sortOrder:
          type: integer
        active:
          type: boolean
...
    Expense:
      type: object
      required: [id, propertyId, title, category, amount, expenseDate, createdAt]
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        propertyName:
          type: ["string", "null"]
        unitId:
          type: ["integer", "null"]
        unitName:
          type: ["string", "null"]
        title:
          type: string
        category:
          type: string
        amount:
          type: number
        expenseDate:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required: [propertyId, title, category, amount, expenseDate]
      properties:
        propertyId:
          type: integer
        unitId:
          type: integer
        title:
          type: string
        category:
          type: string
        amount:
          type: number
        expenseDate:
          type: string
        notes:
          type: string
    ExpenseUpdate:
      type: object
      properties:
        propertyId:
          type: integer
        unitId:
          type: integer
        title:
          type: string
        category:
          type: string
        amount:
          type: number
        expenseDate:
          type: string
        notes:
          type: string
...
    MonthlyFinance:
      type: object
      required: [month, revenue, expenses, netIncome]
      properties:
        month:
          type: string
        revenue:
          type: number
        expenses:
          type: number
        netIncome:
          type: number
...
    FieldUser:
      type: object
      required: [id, name, role, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        propertyId:
          type: integer
          nullable: true
        propertyName:
          type: string
          nullable: true
        status:
          type: string
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
    FieldUserInput:
      type: object
      required: [name, role]
      properties:
        name:
          type: string
        role:
          type: string
        email:
          type: string
        phone:
          type: string
        propertyId:
          type: integer
        status:
          type: string
        notes:
          type: string
...